requirements.txt
datasets==3.6.0
datasketch==1.6.4
Flask==3.0.3
Flask_Cors==4.0.0
jieba==0.42.1
jsonlines==4.0.0
marshmallow==3.22.0
matplotlib==3.5.1
ngrok==1.4.0
nltk==3.8
numpy==1.26.4
openai==1.59.6
peft==0.7.1
psutil==5.9.8
pydantic==2.8.2
rich==13.7.1
scikit_learn==1.5.1
sentence_transformers==2.3.1
simhash==2.1.2
tiktoken==0.10.0
transformers==4.57.6
jinja2==3.1.2
trl==0.13.0
ujson==5.1.0
wandb==0.18.3
streamlit==1.50.0
gradio==5.49.1
modelscope-studio==1.6.1
python-multipart==0.0.22
pyarrow==23.0.0
Pillow==12.0.0
swanlab==0.6.12
soundfile==0.13.1
# torch==2.6.0
# torchvision==0.21.0
